fix(react-native): make HotUpdaterConfig visible to SwiftPM archive target
SignatureVerifier.swift is compiled in the SwiftPM `HotUpdaterArchive`
target, which excludes HotUpdaterImpl.swift. HotUpdaterConfig was defined
inside HotUpdaterImpl.swift, so `swift test` failed with "cannot find
'HotUpdaterConfig' in scope". Extract it into its own HotUpdaterConfig.swift
and add it to the archive target sources. CocoaPods is unaffected (source_files
glob picks up the new file; everything compiles as one module).
